Add check if `column` not in `schema` for `cast_to_datetime`
Allows `cast_to_datetime` to play nicely with the `pl.coalesce` operations when a column doesn't exist 

For example, in Lines 176-183, the column `"birth_datetime"` doesn't exist in the Truven OMOP. Currently, this causes an Exception. Instead, the `cast_to_datetime` call should just return None's and be gracefully handled by the `pl.coalesce` it's called in.